Splits charts and computers into separate packages
Fixes: OX-9673
