+++ src/main/java/sirius/biz/tenants/metrics/charts/AverageEducationPerTenantChart.java
@@ -6,21 +6,19 @@
  * http://www.scireum.de - info@scireum.de
  */
 
-package sirius.biz.tenants.metrics;
+package sirius.biz.tenants.metrics.charts;
 
-import sirius.biz.analytics.events.EventRecorder;
-import sirius.biz.analytics.events.UserActivityEvent;
-import sirius.biz.analytics.events.UserData;
 import sirius.biz.analytics.explorer.ChartFactory;
 import sirius.biz.analytics.explorer.ChartObjectResolver;
-import sirius.biz.analytics.explorer.EventTimeSeriesComputer;
+import sirius.biz.analytics.explorer.MetricTimeSeriesComputer;
 import sirius.biz.analytics.explorer.TimeSeriesChartFactory;
 import sirius.biz.analytics.explorer.TimeSeriesComputer;
 import sirius.biz.jobs.StandardCategories;
 import sirius.biz.tenants.Tenant;
 import sirius.biz.tenants.TenantUserManager;
-import sirius.biz.tenants.Tenants;
 import sirius.biz.tenants.UserAccountController;
+import sirius.biz.tenants.metrics.computers.TenantMetricComputer;
+import sirius.biz.tycho.academy.OnboardingEngine;
 import sirius.kernel.commons.Callback;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
@@ -32,17 +30,22 @@ import javax.annotation.Nullable;
 import java.util.function.Consumer;
 
 /**
- * Uses the {@link sirius.biz.analytics.events.UserActivityEvent} to compute the exact number of active users for a
- * selected period and tenant.
+ * Provides a chart showing the number of {@link TenantMetricComputer#METRIC_AVG_ACTIVITY} for a tenant.
  */
-@Register(framework = Tenants.FRAMEWORK_TENANTS)
-public class NumberOfUserInteractionsPerTenantChart extends TimeSeriesChartFactory<Tenant<?>> {
+@Register
+public class AverageEducationPerTenantChart extends TimeSeriesChartFactory<Tenant<?>> {
 
     @Part
-    private EventRecorder eventRecorder;
+    @Nullable
+    private OnboardingEngine onboardingEngine;
 
     @Override
     public boolean isAccessibleToCurrentUser() {
+        // This is actually the simplest way to determine if either a JDBC or MongoDB based academy is available...
+        if (onboardingEngine == null) {
+            return false;
+        }
+
         UserInfo currentUser = UserContext.getCurrentUser();
         return currentUser.hasPermission(TenantUserManager.PERMISSION_SYSTEM_TENANT_MEMBER)
                || currentUser.hasPermission(UserAccountController.PERMISSION_MANAGE_USER_ACCOUNTS);
@@ -61,7 +64,7 @@ public class NumberOfUserInteractionsPerTenantChart extends TimeSeriesChartFacto
 
     @Override
     protected void collectReferencedCharts(Consumer<Class<? extends ChartFactory<Tenant<?>>>> referenceChartConsumer) {
-        referenceChartConsumer.accept(NumberOfActiveUsersPerTenantChart.class);
+        referenceChartConsumer.accept(AverageEducationPerTenantChart.class);
     }
 
     @Override
@@ -69,24 +72,17 @@ public class NumberOfUserInteractionsPerTenantChart extends TimeSeriesChartFacto
                              boolean hasComparisonPeriod,
                              boolean isComparisonPeriod,
                              Callback<TimeSeriesComputer<Tenant<?>>> executor) throws Exception {
-        EventTimeSeriesComputer<Tenant<?>, UserActivityEvent> computer =
-                new EventTimeSeriesComputer<>(UserActivityEvent.class,
-                                              (tenant, query) -> query.eq(UserActivityEvent.USER_DATA.inner(UserData.TENANT_ID),
-                                                                          tenant.getIdAsString()));
-
-        computer.addAggregation(NumberOfUserInteractionsChart.AGGREGATION_EXPRESSION_USERS, null);
-
-        executor.invoke(computer);
+        executor.invoke(new MetricTimeSeriesComputer<>(TenantMetricComputer.METRIC_AVG_EDUCATION_LEVEL));
     }
 
     @Nonnull
     @Override
     public String getName() {
-        return "TenantNumberOfUserInteractions";
+        return "TenantAvgEducationLevel";
     }
 
     @Override
     public int getPriority() {
-        return 9050;
+        return 9070;
     }
 }
