+++ src/main/java/sirius/biz/tenants/mongo/MongoTenantGlobalMetricComputer.java
@@ -12,7 +12,7 @@ import sirius.biz.analytics.flags.mongo.MongoPerformanceData;
 import sirius.biz.analytics.metrics.mongo.MongoMonthlyGlobalMetricComputer;
 import sirius.biz.model.LoginData;
 import sirius.biz.tenants.UserAccountData;
-import sirius.biz.tenants.metrics.GlobalTenantMetricComputer;
+import sirius.biz.tenants.metrics.computers.GlobalTenantMetricComputer;
 import sirius.db.mongo.Mango;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
